requirements/default.txt
numpy
scipy
pandas
matplotlib
root_numpy
